+++ prefect_dbt/cli/configs/postgres.py
@@ -8,15 +8,15 @@ except ImportError:
 
 from pydantic import Field
 
-from prefect_dbt.cli.configs.base import MissingExtrasRequireError, TargetConfigs
+from prefect_dbt.cli.configs.base import BaseTargetConfigs, MissingExtrasRequireError
 
 try:
-    from prefect_sqlalchemy.credentials import DatabaseCredentials
+    from prefect_sqlalchemy.database import DatabaseCredentials
 except ModuleNotFoundError as e:
     raise MissingExtrasRequireError("Postgres") from e
 
 
-class PostgresTargetConfigs(TargetConfigs):
+class PostgresTargetConfigs(BaseTargetConfigs):
     """
     Target configs contain credentials and
     settings, specific to Postgres.
@@ -63,7 +63,7 @@ class PostgresTargetConfigs(TargetConfigs):
         default="postgres", description="The type of the target."
     )
     credentials: DatabaseCredentials = Field(
-        default_factory=DatabaseCredentials,
+        default=...,
         description=(
             "The credentials to use to authenticate; if there are duplicate keys "
             "between credentials and TargetConfigs, e.g. schema, "